Clean up models and views, added new app for profile and other user-specific things
