+++ pokerole_irl/pokerole_app/models/species_models.py
@@ -1,96 +1,8 @@
 from django.db import models
-from django.contrib.auth.models import User
 
-
-class Type(models.Model):
-    name = models.CharField(max_length=30)
-    resistances = models.ManyToManyField(
-        'self', related_name="not_effective", symmetrical=False, blank=True)
-    weaknesses = models.ManyToManyField(
-        'self', related_name="super_effective", symmetrical=False, blank=True)
-    immunities = models.ManyToManyField(
-        'self', related_name="no_effect", symmetrical=False, blank=True)
-
-    def get_weaknesses_and_resistances(self):
-        weak_res = {}
-        for weakness in self.weaknesses.all():
-            weak_res[weakness] = 2
-        for resistance in self.resistances.all():
-            weak_res[resistance] = 0.5
-        for immunity in self.immunities.all():
-            weak_res[immunity] = 0
-        return weak_res
-
-    def __str__(self):
-        return self.name
-
-    def __repr__(self):
-        return self.__str__()
-
-
-class Ability(models.Model):
-
-    name = models.CharField(max_length=50, default="")
-
-    effect = models.TextField(max_length=1000, default="")
-
-    description = models.TextField(max_length=1000, default="")
-
-    def __str__(self):
-        return self.name
-
-
-class DamageTypeChoices(models.TextChoices):
-    PHYSICAL = "Physical", "Physical"
-    SPECIAL = "Special", "Special"
-    SUPPORT = "Support", "Support"
-
-
-class TargetChoices(models.TextChoices):
-    USER = "User"
-    ONE_ALLY = "One Ally"
-    ALL_ALLIES = "User & All Allies in Range"
-    FOE = "Foe"
-    RANDOM_FOE = "Random Foe"
-    ALL_FOES = "All Foes in Range"
-    AREA = "Area"
-    BATTLEFIELD = "Battlefield"
-
-
-class Move(models.Model):
-
-    name = models.CharField(max_length=50)
-
-    move_type = models.ForeignKey(Type, related_name="moves",
-                                  on_delete=models.PROTECT, null=True, blank=True)
-
-    power = models.IntegerField(null=True, blank=True)
-
-    damage_stat = models.CharField(max_length=30, null=True, blank=True)
-
-    damage_modifier = models.CharField(max_length=30, null=True, blank=True)
-
-    primary_accuracy = models.CharField(max_length=30, null=True, blank=True)
-
-    secondary_accuracy = models.CharField(max_length=30, null=True, blank=True)
-
-    description = models.TextField(
-        max_length=1000, default="", null=True, blank=True)
-
-    effect = models.TextField(max_length=1000, null=True, blank=True)
-
-    attributes = models.JSONField(default=dict)
-
-    added_effects = models.JSONField(default=dict)
-
-    damage_type = models.CharField(
-        max_length=8, choices=DamageTypeChoices.choices, null=True, blank=True)
-
-    target = models.CharField(
-        max_length=30, choices=TargetChoices.choices, null=True, blank=True)
-
-    def __str__(self):
-        return self.name
+from .base_models import Type
+from .ability_models import Ability
+from .move_models import Move
 
 
 class PokemonSpecies(models.Model):
@@ -229,62 +141,3 @@ class MoveSet(models.Model):
     species = models.ForeignKey(PokemonSpecies, on_delete=models.CASCADE)
     move = models.ForeignKey(Move, on_delete=models.CASCADE)
     learned = models.CharField(max_length=30)
-
-
-class Pokedex(models.Model):
-    name = models.CharField(max_length=255)
-    species = models.ManyToManyField(PokemonSpecies, through='PokedexEntry')
-
-    def __str__(self):
-        return self.name
-
-
-class PokedexEntry(models.Model):
-    species = models.ForeignKey(
-        PokemonSpecies, on_delete=models.CASCADE, related_name="entries")
-    pokedex = models.ForeignKey(Pokedex, on_delete=models.CASCADE)
-    number = models.IntegerField()
-
-    RARITIES = models.TextChoices(
-        "Rarities", "Legendary Very_rare Rare Uncommon Common")
-    rarity = models.CharField(
-        max_length=30, choices=RARITIES.choices, null=True, blank=True)
-
-    def __str__(self):
-        return f"{self.pokedex} #{self.number}: {self.species}"
-
-
-class Nature(models.Model):
-    name = models.CharField(max_length=30)
-    confidence = models.PositiveIntegerField()
-    keywords = models.CharField(max_length=50, default="")
-    description = models.TextField(max_length=1000, default="")
-
-    def __str__(self):
-        return self.name
-
-
-class Item(models.Model):
-    name = models.CharField(max_length=30)
-    description = models.TextField(max_length=1000, default="")
-    type_bonus = models.ForeignKey(
-        Type, blank=True, null=True, on_delete=models.SET_NULL)
-    value = models.CharField(max_length=50, default="")
-    species = models.ForeignKey(
-        PokemonSpecies, blank=True, null=True, on_delete=models.CASCADE)
-    heal_amount = models.CharField(max_length=30, default="")
-    suggested_price = models.CharField(max_length=30, default="")
-    PMD_price = models.PositiveIntegerField(blank=True, null=True)
-
-    def __str__(self):
-        return self.name
-        
-class Profile(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-
-    #avatar = models.ImageField(default='default.jpg', upload_to='profile_images')
-    #Will be added later
-    bio = models.TextField()
-
-    def __str__(self):
-        return self.user.username
