+++ pokerole_irl/pokerole_irl/urls.py
@@ -20,6 +20,6 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('pokerole_app.urls')),
+    path('accounts/', include('user_app.urls')),
     path('', include('django.contrib.auth.urls')),
-    path('health-check/', include('health_check.urls'))
 ] + staticfiles_urlpatterns()
