+++ pokerole_irl/user_app/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import Profile
+
+admin.site.register(Profile)
