+++ pokerole_irl/user_app/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from .views import RegisterView, UserProfileView
+
+urlpatterns = [
+    path('register/', RegisterView.as_view(), name='register'),
+    path('profile/', UserProfileView.as_view(), name='user-profile')
+]
